project/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
huggingface-hub==0.17.3
huggingsound==0.1.6
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jiwer==2.6.0
joblib==1.3.2
librosa==0.9.2
llvmlite==0.41.0
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
multiprocess==0.70.15
numba==0.58.0
numpy==1.24.4
packaging==23.2
pandas==2.0.3
platformdirs==3.10.0
pooch==1.7.0
pyarrow==13.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.6
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==2.13.7
regex==2023.8.8
requests==2.31.0
resampy==0.4.2
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
starlette==0.27.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==1.12.1
tqdm==4.66.1
transformers==4.33.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
